requirements.txt
Flask==1.1.1
requests==2.22.0
waitress==1.4.3
validate-email==1.3
mysql-connector-python==8.0.19